bluetooth: btusb: Always try alt 1 for transparent SCO aka WBS aka mSBC
It seems many USB adapters don't support the recommended alt setting 6.
But they will work with alt 1 or possibly other alt settings.

This is used for transparent SCO, which is in turn used to pass already
compressed audio to/from the adapter.  As opposed to normal SCO, which
passes uncompressed audio that will be compressed by the adapter.

Transparent SCO is normally used for wide band speech (WBS), which uses
the mSBC codec at 16 kHz.  The recommended packet size for alt 6 is well
suited for the mSBC packet size used in WBS.

This driver's alt setting logic assumes eSCO will only be used for WBS,
and tries to optimize trasnsmission for that, which isn't necessarily
true.  It also assumes the endpoint's alt modes are exactly as
recommended in the bt spec.

It seems like it would be better if the socket user could say, "I want
to send a 60 byte packet every 7.5 ms." And then the btusb driver could
choose the best way to do that.